Hierarchical squarified treemap works.

+++ include/Treemap.h
@@ -1,6 +1,7 @@
 #ifndef TREEMAP_H
 #define TREEMAP_H
 
+#include <iostream>
 #include <vector>
 #include <cfloat>
 #include "math.h"
@@ -15,8 +16,8 @@ class Treemap
 {
 public:
 	Treemap(EntityTree et, int _width, int _height);
-	void treemapMultidimensional(vector<BaseEntity*> *data, int width, int height, int xOffset, int yOffset);
-	void treemapSingledimensional(vector<BaseEntity*> *data, int width, int height, int xOffset, int yOffset);
+	void treemapMultidimensional(vector<BaseEntity*> *data, float width, float height, float xOffset, float yOffset);
+	void treemapSingledimensional(vector<BaseEntity*> *data, float width, float height, float xOffset, float yOffset);
 	void squarify(vector<BaseEntity*> *data, vector<BaseEntity*> *currentRow, Container container);
 	void normalize(vector<BaseEntity*> *data, int area);
 	int sumScores(vector<BaseEntity*> *data);
